Calculate last dose amount (LDOS) and time after last dose (TAD)

Use lastdose to add (or potentially replace) columns to the input data frame; lastdose_list and lastdose_df returns calculated information as either list or data.frame format without modifying the input data.

lastdose(data, ..., include_ldos = TRUE)

lastdose_list(
  data,
  fill = -99,
  back_calc = TRUE,
  addl_ties = c("obs_first", "dose_first"),
  comments = find_comments(data)
)

lastdose_df(data, ...)

Arguments

data
data set as data frame; see details

...
arguments passed to lastdose_list

include_ldos
logical; if FALSE then the LDOS data is not appended to the data set. Only used for the lastdose function.

fill
the value for TAD that is used for records when no doses are found for an individual or when back_calc is FALSE.

back_calc
if TRUE, then the time before the first dose is calculated for records prior to the first dosing record when at least one dosing record is found in the data set. Records before the first dosing record will have negative values.

addl_ties
what to do when doses scheduled through ADDL happen at the same time as observation records; if obs_first then the observation is assumed to happen before the dose and the observation is a trough concentration; if dose_first then the dose is assumed to be administered and the observation made immediately after (with no advance in time). See details.

comments
a logcial vector with length equal to the number of rows in data indicating which records are to be ignored when looking for TAD and LDOS.

Details

When calling lastdose to modify the data frame, two columns will be added (by default): TAD indicating the time after the most-recent dose and LDOS indicating the amount of the most recent dose. This default behavior can be modified with the include_ldos argument.

When calling lastdose_list or lastdose_df, the respective items are accessible with tad and ldos (note the lower case form here to distinguish from the columns that might be added to the data frame).

Additional notes regarding behavior:

All functions require an input data set as a data frame

The data set should be formatted according to NMTRAN type conventions

Required columns

ID or id: subject identifier

TIME or time: data set time

AMT or amt: dose amount for dosing records

EVID or evid: event ID; records with EVID or 1 or 4 are considered dosing records

Optional columns

ADDL or addl: additional doses to administer

II or ii: dosing interval

An error is generated if required columns are not found; no error or warning if optional columns are not found

All required and optional columns are required to be numeric

Missing values are not allowed in: ID,EVID,ADDL,II

When missing values are found in TIME, both TAD and LDOS are set to missing

An error is generated for missing AMT in dosing records (evid 1 or 4)

No error is generated for missing AMT in non-dosing records

An example illustrating the addl_ties argument: when there is Q24h dosing and both an an additional dose and an observation happen at 24 hours, obs_first will set the observationTAD to 24 and dose_first will set the observation TAD to 0.

Examples

file <- system.file("csv/data1.csv", package="lastdose")

require("Rcpp")

#> Loading required package: Rcpp

data <- read.csv(file)

a <- lastdose(data)

b <- lastdose_df(data)

c <- lastdose_list(data)